requirements.txt
anthropic
jira
PyGithub
python-dotenv
requests
